Restructure node-sass plugin
