+++ src/plugins/node-sass/index.ts
@@ -1,0 +1,2 @@
+export { compile } from "./compile";
+export { filterSass } from "./filterSass";
